a function that sets the value of a bit to 1 at a given index.

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -15,6 +15,6 @@ int set_bit(unsigned long int *n, unsigned int index)
 		return (-1);
 	}
 
-	*n = *n | ((unsigned long)1 << index);
+	*n = *n | (1 << index);
 	return (1);
 }
